refactor: add ui branching flag to components
